rework static page resolvemenuitem
