+++ classes/MLStaticPage.php
@@ -1,13 +1,5 @@
 <?php namespace Winter\Translate\Classes;
 
-use Url;
-
-use Winter\Storm\Router\Router;
-
-use Winter\Pages\Classes\Page as StaticPage;
-
-use Winter\Translate\Models\Locale;
-
 /**
  * Represents a multi-lingual Static Page object.
  *
@@ -109,69 +101,5 @@ class MLStaticPage extends MLCmsObject
      */
     protected function checkSafeMode()
     {
-    }
-
-    protected static function resolveMenuItem($item, $url, $theme)
-    {
-        $result = [];
-        $locales = Locale::listEnabled();
-
-        if ($item->type == 'static-page') {
-            if (!$item->reference) {
-                return;
-            }
-            $page = StaticPage::find($item->reference);
-            $alternateLinks = [];
-            foreach ($locales as $locale => $name) {
-                if ($pageUrl = static::getMLPageUrl($page, $locale)) {
-                    $alternateLinks[$locale] = Url::to($pageUrl);
-                }
-            }
-
-            foreach ($alternateLinks as $locale => $link) {
-                $result[] = [ 
-                    'url' => $link,
-                    'mtime' => $page->mtime,
-                    'alternateLinks' => $alternateLinks,
-                ];
-            }
-            return $result;
-
-        } elseif ($item->type == 'all-static-pages') {
-            if (empty($pages = StaticPage::all())) {
-                return;
-            }
-            $recordItems = [];
-            foreach ($pages as $page) {
-                $alternateLinks = [];
-                foreach ($locales as $locale => $name) {
-                    if ($pageUrl = static::getMLPageUrl($page, $locale)) {
-                        $alternateLinks[$locale] = Url::to($pageUrl);
-                    }
-                }
-
-                foreach ($alternateLinks as $locale => $link) {
-                    $recordItems['items'][] = [ 
-                        'url' => $link,
-                        'mtime' => $page->mtime,
-                        'alternateLinks' => $alternateLinks,
-                    ];
-                }
-            }
-            $result[] = $recordItems;
-        }
-
-        return $result;
-    }
-
-    protected static function getMLPageUrl($page, $locale)
-    {
-        $translator = Translator::instance();
-        $translator->setLocale($locale);
-
-        $page->rewriteTranslatablePageUrl($locale);
-        $url = $translator->getPathInLocale(array_get($page->attributes, 'viewBag.url'), $locale);
-
-        return (new Router)->urlFromPattern($url);
     }
 }
